Update SCons (#17278)
Closes #17254

Summary of the issue:
We are currently on SCons version 4.5.2, which was released in March of 2023. The latest version of SCons, version 4.8.1, was released in September 2024.

Description of user facing changes:
None.

Description of development approach
* Updated the required version of SCons from 4.5.2 to 4.8.1 in `requirements.txt`.
* Changed the order of arguments to  `SCons.Tool.MSCommon.vc.find_vc_pdir` to that expected by SCons 4.8.
  * Previously, the order of arguments was `environment, version`. As of SCons 4.8, it is `version, environment`.
  * Changed in SCons/scons#4534

Testing strategy:
Built the project locally and on Appveyor.

Known issues with pull request:
None.