Add ability to speak next and previous lines/paragraphs when navigating with braille (#17256)
Fixes #17053

Summary of the issue:
Sometimes, it may be desirable tohear lines when navigating with braille.

Description of user facing changes
A new command, not assigned to gestures, is available in the braille category to toggle speaking of lines/paragraphs when navigating with braille, to enable or disable this feature. A new checkbox for this has been added in the braille settings panel as well.
Description of development approach
A new key for this feature has been added to configSpec.py. The value is a boolean (False by default).
A new script has been added to globalCommands.py.
A new function has been added to braille.py, to be used by next and previousLine functions, also modified so that info is expanded to the corresponding read unit, i.e., line or paragraph.
The braille settings panel has a new checkbox to toggle this feature.