Fixed text for allow NVDA to control the volume of other apps (#17317)
Fixes #17303

Summary of the issue:
Associated to the audio new option "Allow NVDA to control the volume of other applications" (ex Volume adjuster), there is a command (script) with no default bound gesture to toggle the value of this parameter. There are two issues with this script:

Using it only reports "Yes" or "No", whereas other toggle scripts are usually more precise, e.g. the command "Toggles on and off the inclusion of layout tables in browse mode" reports "layout tables off" or "layout tables on".
The input help or command name in the input gestures dialog still mentions volume adjuster, whereas the option in the audio settings and in the User Guide has been reworded.
Description of user facing changes
Reworded the command name / input help.
Calling the command now reports something more explicit.