Report destination of link in Word / Outlook legacy (#17292)
Summary of the issue:
The command to report the destination of a link (NVDA+k) does not work in Word or Outlook when using object model (by default for Windows 10 users or users of older Office versions).

Description of user facing changes
NVDA+k will wow report the destination of the link in Word/Outlook object model.

Description of development approach
Created a specific _getLinkDataAtCaretPosition function to report the required information on a link in a WordDocumentTextInfo. Other text info classes can define such functions in the future to support the command in other content types.
I have not moved the initial implementation to support link reporting in browsers since I do not know exactly in which text info I should move it to guarantee that I'll not break any already supported use case. I may change move this if requested.