Improve fetching of document layout information in word (#17340)
Summary of the issue:
If the Word document description is None, None is reported after the document layout.

Description of user facing changes
The document layout Will be reported without garbage information like None.

Description of development approach
Check if the description is an instance of str, and if it's not just blank characters. In this case, prepend the document layout. Otherwise, just use the document layout as the description.