Add options to set translation and input braille tables according to NVDA's language (#17314)
Fixes issue #17306

Summary of the issue:
By default, NVDA sets braille tables according to its language. When NVDA's language is changed, by default braille tables are modified accordingly, which may be confusing for users.

Description of user facing changes
In the Braille settings panel, combo boxes for tables have an option to set them according to NVDA's language, so that now this is optional. The automatically selected tables corresponding to the current language are described in parentheses.

Description of development approach
New options have been added to the Braille settings panel. braille, brailleInput and brailleTables modules have been modified so that the "auto" option in configuration can be optional.