Only warn about add-on compatibility when updating to an incompatible API (#17370)
Closes #17071 

Summary of the issue:
NVDA complains about incompatible updates whenever it is updated with add-ons installed that are incompatible with the updated API version. This is particularly annoying for beta and alpha users.

Description of user facing changes
The add-on incompatibility message is now only presented when updating to a version with an incompatible API to that of the installed copy. That is, when the update will cause `addonAPIVersion.BACK_COMPAT_TO[0]` to increase.

Description of development approach
Make increasing the add-on API backcompat's year a precondition of showing the warning in `updateCheck.py`.

Testing strategy:
Created two builds of NVDA, both with their version spoofed to 2024.4, and update version type to alpha.
In the first, set `addonAPIVersion.BACK_COMPAT_TO` to `(2024.1.0)`, and the other to `(2025.1.0)`.
Installed an add-on that is incompatible with the 2025 API in both copies.
Ensured that updating the first copy caused the add-on incompatibility message to appear, and updating the second copy did not.

Known issues with pull request:
None.