Fix bug to move braille to next line in UIA documents (#17401)
Link to issue number:
Fixes #17251

Summary of the issue:
In some UIA documents which end with an empty line, braille cannot be moved to the last line.

Description of user facing changes
Braille can be moved to the last line in all UIA documents.
If the braille move to next line command is run from the last line, the cursor will be moved to the last character. In this case, if the review cursor follows system cursor, the review cursor will also be moved.

Description of development approach
A `shouldCollapseToEnd` variable has been added to the method to move braille to the next line, to determine if text info should be collapsed to end. When braille is not moved to the next line using other procedures, `shouldCollapseToEnd` is set to True and text info is collapsed to end.

Testing strategy:
Tested manually in Notepad.