Add commands to move the review cursor to the first and last character of selected text (#17393)
Link to issue number:
Fixes #17299

Summary of the issue:
NVDA doesn't have commands to move the review cursor to start and end of selection, which may be useful to review the selected text.

Description of user facing changes
`NVDA+alt+home` and `NVDA+alt+end` can be used to move the review cursor to the first and last carácter of selected text.

Description of development approach
In globalCommands.py, GlobalCommands class, a staticmethod has been added to get the current selection. This is used to scripts added to move the review cursor to start and end of selection, similar to other scripts used to move the review cursor.

Testing strategy:
Tested manually in a webpage and in Notepad.