Fix braille next line command for documents of Notepad++ (#17431)
Fixes #17430

Summary of the issue:
Next braille line command doesn't work in Notepad++ documents when the last line is empty.

Description of user facing changes
Next braille line command works as expected in Notepad++ documents.

Description of development approach
The collapse method is overriden for Scintilla textInfo, so that textInfo is expanded to line before collapsing to end. In this way the cursor is properly moved to the end of the document when running the braille next line command.