PowerPoint: Report link destination (#17435)
When in PowerPoint, NVDA+K does not work to report a link.

Description of user facing changes
In PowerPoint slides (edition mode), pressing NVDA+K will now report the link destination. This works in text areas as well as on shapes / charts / images that are links.
Also 2 small fixes:
Restored the Copy and Close buttons of the browseable message of NVDA+K that were removed by mistake in Report destination of link in Word / Outlook legacy #17292
In Windows UI, e.g. settings, pressing NVDA+K on a link does not report anymore that there is no link; it rather reports that the destination cannot be reported.
Description of development approach
Use PowerPoint object model
Changed the global strategy to retrieve a link: first try from text infos; if not supported, try from the focused object.