Removed a reference to the launcher audio logo from the quick start guide (#17519)
Summary of the issue:

In #17507, the audio logo played while the NVDA launcher is loading has been removed. However, the quick start guide still makes a reference to the music playing before a temporary copy of NVDA loads.

Description of user facing changes

The quick start guide no longer includes misleading information. Instead, it has been replaced with a note that it may take a few seconds for the temporary copy to load.


---------

Co-authored-by: Sean Budd <seanbudd123@gmail.com>