Handle BrailliantB devices that do not report their number of cells (#17526)
Partially fixes #17518

Summary of the issue:
Some HumanWare devices do not report the number of cells in their HID capabilities report. This is the case for Brailliant BI 40X since firmware 2.4.

Description of user facing changes
The Brailliant BI 40X and similar display with firmware version 2.4 is working correctly again.

Description of development approach
This PR fixes the issue by using the HID output report size to calculate the number of cells. If the device reports a cell count, this is still being used.

Testing strategy:
Tested with a Brailliant BI 40X over Bluetooth.

Known issues with pull request:
Since firmware 2.4 this device is recognized by the Standard HID driver over Bluetooth. However, the keys do not work using that driver. This PR does not fix the Standard HID driver.