Fix math not being read in Chromium-based browsers (#17421) (#17548)
Fixes #17421

Summary of the issue:
Chromium-browsers were changed to return E_NOTIMPL and broke speech and braille of all math equations (i.e., no speech or braille for math) in those browsers because NVDA passes along a 'not implemented' COM exception.

Description of user facing changes
This restores the behavior of NVDA (i.e., math reads) as before the change.

Description of development approach
A try/except block is added around the call. If the error is E_NOTIMPL, the code moves on as before. Otherwise, the error is re-raised as before.