Add more notes on triage (#17550)
To improve objectivity in the triage process, community members are discouraged from self-triaging their own issues. As employees, we generally do not triage our own issues either.
To prevent interfering in NV Access's workflow, clarifications around closing tickets need to be improved.
Tickets which are pending a response from NV Access should not be closed, to help ensure a response is given.