Fix braille table loading with debug logging enabled and/or add-ons disabled (#17565)
Summary of the issue:
When custom braille tables are available in the manifest of the scratchpad and NVDA starts with debug logging enabled, it will raise an unbound local error due to a wrong logging statement. This actually revealed another bug, namely that scratchpad braille tables weren't ignored when addons are disabled.

Description of user facing changes
No longer an error in this specific case. Furthermore, custom braille tables from the scratchpad wil be disabled when addons are.

Description of development approach
Fixed the debug logging statement
in brailleTables.initialize, return early when add-ons disabled or when appX.