Add ability to add more constraints to bdDetect USB device registration (#17537)
Fixes #17521

Summary of the issue:
In several cases, registering a USB device in bdDetect with just VID and PID is not enough. We need an extra filter to ensure that only the right devices are detected. This was fixed at the driver level until now, but fixing this in bdDetect improves performance and stability.

Description of user facing changes
More stability and quicker detection times.

Description of development approach
Added the ability to register a match func that further constraints device detection for specific VID/PID combinations.