Remove problematic euro symbol in fi (#17578)
Link to issue number:

#17471

Summary of the issue:

Euro complexSymbol in fi locale causing euro currency numbers to not be read in microsoft excel.

Description of user facing changes

Affects the reading of euro currency numbers in microsoft excel. The behavior should be in line with that of some common locales such as english and spanish where this regex is not present.

Description of development approach

Followed the advice of seanbudd and removed the regex, after spot-checking to verify that this is likely a finnish-only regex and behavior.