Fixes #17418)
Summary of the issue:
Nvda crashes when selecting whole text in specific source files in Android Studio and IntelliJ Idea

Description of user facing changes
NVDA should no longer crash.

Description of development approach
Java Access Bridge has a buffer limit of 10240 characters for the text. If the text is too large to fit on this buffer size, NVDA will split the calls to take the buffer size in account.