Make continuous reading work when using SAPI5 voices without bookmark support (#17523)
Fixes #16691

Summary of the issue:
Some SAPI5 synthesizers did not work correctly with continuous reading, stopping speech at the end of a sentence, because they don't properly support bookmark events.

Description of user facing changes
Some of the problematic SAPI5 synthesizers will be able to work correctly with continuous reading.

Description of development approach
NVDA relies on bookmark events, which have to be implemented by the TTS engine itself. But no matter what events the engine supports, StartStream and EndStream events are always generated by the SAPI5 framework.

This fix does the following.

SAPI5 fix
When requested to speak a new sequence, store the bookmarks in a list (deque), and put the bookmark list into a dict _streamBookmarksNew with the stream number as its key. This dict stores bookmark lists before the stream actually starts, because the previous stream (which may use the same stream number) may not have been ended when self.tts.Speak returns.

When StartStream event is raised, move the bookmark list of this stream from _streamBookmarksNew to _streamBookmarks, which stores bookmark lists of currently speaking streams.

When EndStream event is raised, bookmarks stored for the current stream will be triggered (through synthIndexReached) before synthDoneSpeaking. This way NVDA will continue to the next line instead of waiting indefinitely.

As most synthesizers do support bookmarks, to prevent triggering the same bookmark twice, the bookmark stored in the list will be removed when the Bookmark event for it is raised.

Thread-safety: self.tts.Speak and TTS events are handled in the same thread, so there should be no multithread-related problems.

SAPI4 fix
Similar to the SAPI5 fix, which stores bookmarks and replay them when AudioStop is raised.

However, SAPI4 does not provide "stream numbers" to distinguish each request, yet SAPI4 supports queueing speech requests. The fix assumes that speech requests are always processed in a first-in, first-out manner, and uses a deque _bookmarkLists to store bookmark lists from different requests.

When cancel is called, the fix assumes that all queued speech requests are cancelled immediately, and clears the bookmark lists.